Refactor `FormValidator`: move extension functions to class members
