+++ kmm-form-validation/src/commonMain/kotlin/ru/mobileup/kmm_form_validation/validation/form/FormValidatorDsl.kt
@@ -62,7 +62,7 @@ class FormValidatorBuilder {
     }
 
     fun build(coroutineScope: CoroutineScope): FormValidator {
-        return FormValidator(validators).apply {
+        return FormValidator(validators, coroutineScope).apply {
             features.forEach { feature ->
                 feature.install(coroutineScope, this)
             }
